pyproject.toml
[tool.poetry]
name = "chatbot-project"
version = "0.1.0"
description = "A simple chatbot with Django DRF and Google Studio API"
authors = ["John Raphy Karippery <33422429+karippery@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.8"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
environ = "^1.0"
django-cors-headers = "^4.9.0"
drf-spectacular = "^0.29.0"
whitenoise = "^6.11.0"
google-genai = "^1.52.0"

[tool.poetry.group.dev.dependencies]
black = "^25.11.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
pytest = "^9.0.1"
pytest-django = "^4.11.1"
django-debug-toolbar = "^6.1.0"
django-extensions = "^4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"